+++ packages/terra-data-grid/tests/jest/DataGrid.test.jsx
@@ -92,7 +92,7 @@ describe('DataGrid', () => {
   it('verifies the rows are created with the right props', () => {
     const verifyRow = (rowIndex, rowComponent, data, overflowColumns) => {
       expect(rowComponent.props.displayedColumns).toEqual(overflowColumns);
-      expect(rowComponent.props.hasRowSelection).toBe(false);
+      expect(rowComponent.props.rowSelectionMode).toBeUndefined();
       expect(rowComponent.key).toEqual(data.id);
       expect(rowComponent.props.onCellSelect).toBeDefined();
       expect(rowComponent.props.rowHeaderIndex).toEqual(0);