+++ packages/terra-framework-docs/src/terra-dev-site/test/table/TableSingleRowSelection.test.jsx
@@ -10,6 +10,7 @@ const tableDataJSON = {
   rows: [
     {
       id: '1',
+      isSelectable: true,
       cells: [
         { content: 'Fleck, Arthur' },
         { content: '1007-MTN' },
@@ -18,6 +19,7 @@ const tableDataJSON = {
     },
     {
       id: '2',
+      isSelectable: true,
       cells: [
         { content: 'Wayne, Bruce' },
         { content: '1107-MTN-DR' },
@@ -27,6 +29,7 @@ const tableDataJSON = {
     },
     {
       id: '3',
+      isSelectable: true,
       cells: [
         { content: 'Carr, Alicia' },
         { content: '1007-MTN' },
@@ -36,6 +39,7 @@ const tableDataJSON = {
     },
     {
       id: '4',
+      isSelectable: true,
       cells: [
         { content: 'McMurphy, Donald' },
         { content: '1024-MTN', isMasked: true },
@@ -44,6 +48,7 @@ const tableDataJSON = {
     },
     {
       id: '5',
+      isSelectable: true,
       cells: [
         { content: 'Peters, Timothy' },
         { content: '1207-MTN' },
@@ -53,39 +58,46 @@ const tableDataJSON = {
   ],
 };
 
-// eslint-disable-next-line compat/compat
-const params = new URLSearchParams(window.location.search);
-const rowSelectionHeaderSelectable = params ? params.get('header-select') : false;
-
-const TableRowSelection = () => {
+const TableSingleRowSelection = () => {
+  const rowHeaderIndex = 0;
   const { cols, rows } = tableDataJSON;
   const [rowData, setRowData] = useState(rows);
 
   const onRowSelect = useCallback((rowId) => {
-    // Remove current selections
     const newRowData = [...rowData];
 
     const dataRowToUpdate = newRowData.find(row => row.id === rowId);
-    if (dataRowToUpdate) {
-      dataRowToUpdate.isSelected = !dataRowToUpdate.isSelected;
-    }
+
+    newRowData.forEach((row) => {
+      if (row.id !== dataRowToUpdate.id) {
+        // eslint-disable-next-line no-return-assign, no-param-reassign
+        row.isSelected = false;
+      }
+    });
+
+    dataRowToUpdate.isSelected = !dataRowToUpdate.isSelected;
 
     setRowData(newRowData);
   }, [rowData]);
 
+  const onRowSelectionHeaderSelect = useCallback(() => {
+    // eslint-disable-next-line no-alert
+    alert('Row Selection Header Clicked');
+  }, []);
+
   return (
     <Table
-      id="table-with-row-selections"
+      id="table-with-single-row-selection"
       overflowColumns={cols}
       rows={rowData}
-      rowHeaderIndex={1}
+      rowHeaderIndex={rowHeaderIndex}
       columnWidth="180px"
-      ariaLabel="Table with Row Selections"
-      rowSelectionMode="multiple"
-      onRowSelect={onRowSelect}
-      onRowSelectionHeaderSelect={rowSelectionHeaderSelectable ? () => {} : undefined}
+      ariaLabel="Table with Single Row Selection"
+      rowSelectionMode="single" // Prop to turn row selection mode on/off
+      onRowSelect={onRowSelect} // For row selection, consumer must provide a callback that the Worklist Data Grid will call when the user selects one or more rows.
+      onRowSelectionHeaderSelect={onRowSelectionHeaderSelect}
     />
   );
 };
 
-export default TableRowSelection;
+export default TableSingleRowSelection;