+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added examples and tests to `terra-table` for single row selection mode.
+
 ## 1.45.0 - (November 20, 2023)
 
 * Added